refactor: CategoryEntity를 Notice domain 페키지로 이동
