+++ src/main/java/com/kustacks/kuring/notice/domain/NoticeRepository.java
@@ -1,6 +1,5 @@
 package com.kustacks.kuring.notice.domain;
 
-import com.kustacks.kuring.category.domain.CategoryName;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
