+++ src/main/java/com/kustacks/kuring/user/facade/UserQueryFacade.java
@@ -1,6 +1,6 @@
 package com.kustacks.kuring.user.facade;
 
-import com.kustacks.kuring.category.domain.CategoryName;
+import com.kustacks.kuring.notice.domain.CategoryName;
 import com.kustacks.kuring.common.firebase.FirebaseService;
 import com.kustacks.kuring.notice.common.dto.CategoryNameDto;
 import com.kustacks.kuring.notice.common.dto.DepartmentNameDto;
