+++ src/main/java/com/kustacks/kuring/worker/update/notice/CategoryNoticeUpdater.java
@@ -1,6 +1,6 @@
 package com.kustacks.kuring.worker.update.notice;
 
-import com.kustacks.kuring.category.domain.CategoryName;
+import com.kustacks.kuring.notice.domain.CategoryName;
 import com.kustacks.kuring.common.firebase.FirebaseService;
 import com.kustacks.kuring.notice.domain.Notice;
 import com.kustacks.kuring.notice.domain.NoticeRepository;
