+++ src/test/java/com/kustacks/kuring/user/domain/UserTest.java
@@ -1,6 +1,6 @@
 package com.kustacks.kuring.user.domain;
 
-import com.kustacks.kuring.category.domain.CategoryName;
+import com.kustacks.kuring.notice.domain.CategoryName;
 import com.kustacks.kuring.notice.domain.DepartmentName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
